+++ server/events/event_parser_test.go
@@ -42,7 +42,10 @@ func TestParseGithubRepo(t *testing.T) {
 		CloneURL:          "https://github-user:github-token@github.com/owner/repo.git",
 		SanitizedCloneURL: Repo.GetCloneURL(),
 		Name:              "repo",
-		Hostname:          "github.com",
+		VCSHost: models.VCSHost{
+			Hostname: "github.com",
+			Type:     models.Github,
+		},
 	}, r)
 }
 
@@ -88,7 +91,10 @@ func TestParseGithubIssueCommentEvent(t *testing.T) {
 		CloneURL:          "https://github-user:github-token@github.com/owner/repo.git",
 		SanitizedCloneURL: *comment.Repo.CloneURL,
 		Name:              "repo",
-		Hostname:          "github.com",
+		VCSHost: models.VCSHost{
+			Hostname: "github.com",
+			Type:     models.Github,
+		},
 	}, repo)
 	Equals(t, models.User{
 		Username: *comment.Comment.User.Login,
@@ -122,7 +128,7 @@ func TestParseGithubPull(t *testing.T) {
 	_, _, err = parser.ParseGithubPull(&testPull)
 	ErrEquals(t, "number is null", err)
 
-	pullRes, repoRes, err := parser.ParseGithubPull(&Pull)
+	pullRes, _, err := parser.ParseGithubPull(&Pull)
 	Ok(t, err)
 	Equals(t, models.PullRequest{
 		URL:        Pull.GetHTMLURL(),
@@ -131,16 +137,18 @@ func TestParseGithubPull(t *testing.T) {
 		HeadCommit: Pull.Head.GetSHA(),
 		Num:        Pull.GetNumber(),
 		State:      models.Open,
+		BaseRepo: models.Repo{
+			Owner:             "owner",
+			FullName:          "owner/repo",
+			CloneURL:          "https://github-user:github-token@github.com/owner/repo.git",
+			SanitizedCloneURL: Repo.GetCloneURL(),
+			Name:              "repo",
+			VCSHost: models.VCSHost{
+				Hostname: "github.com",
+				Type:     models.Github,
+			},
+		},
 	}, pullRes)
-
-	Equals(t, models.Repo{
-		Owner:             "owner",
-		FullName:          "owner/repo",
-		CloneURL:          "https://github-user:github-token@github.com/owner/repo.git",
-		SanitizedCloneURL: Repo.GetCloneURL(),
-		Name:              "repo",
-		Hostname:          "github.com",
-	}, repoRes)
 }
 
 func TestParseGitlabMergeEvent(t *testing.T) {
@@ -150,6 +158,19 @@ func TestParseGitlabMergeEvent(t *testing.T) {
 	Ok(t, err)
 	pull, repo, err := parser.ParseGitlabMergeEvent(*event)
 	Ok(t, err)
+
+	expRepo := models.Repo{
+		FullName:          "gitlabhq/gitlab-test",
+		Name:              "gitlab-test",
+		SanitizedCloneURL: "https://example.com/gitlabhq/gitlab-test.git",
+		Owner:             "gitlabhq",
+		CloneURL:          "https://gitlab-user:gitlab-token@example.com/gitlabhq/gitlab-test.git",
+		VCSHost: models.VCSHost{
+			Hostname: "example.com",
+			Type:     models.Gitlab,
+		},
+	}
+
 	Equals(t, models.PullRequest{
 		URL:        "http://example.com/diaspora/merge_requests/1",
 		Author:     "root",
@@ -157,16 +178,10 @@ func TestParseGitlabMergeEvent(t *testing.T) {
 		HeadCommit: "da1560886d4f094c3e6c9ef40349f7d38b5d27d7",
 		Branch:     "ms-viewport",
 		State:      models.Open,
+		BaseRepo:   expRepo,
 	}, pull)
 
-	Equals(t, models.Repo{
-		FullName:          "gitlabhq/gitlab-test",
-		Name:              "gitlab-test",
-		SanitizedCloneURL: "https://example.com/gitlabhq/gitlab-test.git",
-		Owner:             "gitlabhq",
-		CloneURL:          "https://gitlab-user:gitlab-token@example.com/gitlabhq/gitlab-test.git",
-		Hostname:          "example.com",
-	}, repo)
+	Equals(t, expRepo, repo)
 
 	t.Log("If the state is closed, should set field correctly.")
 	event.ObjectAttributes.State = "closed"
@@ -180,7 +195,18 @@ func TestParseGitlabMergeRequest(t *testing.T) {
 	var event *gitlab.MergeRequest
 	err := json.Unmarshal([]byte(mergeRequestJSON), &event)
 	Ok(t, err)
-	pull := parser.ParseGitlabMergeRequest(event)
+	repo := models.Repo{
+		FullName:          "gitlabhq/gitlab-test",
+		Name:              "gitlab-test",
+		SanitizedCloneURL: "https://example.com/gitlabhq/gitlab-test.git",
+		Owner:             "gitlabhq",
+		CloneURL:          "https://gitlab-user:gitlab-token@example.com/gitlabhq/gitlab-test.git",
+		VCSHost: models.VCSHost{
+			Hostname: "example.com",
+			Type:     models.Gitlab,
+		},
+	}
+	pull := parser.ParseGitlabMergeRequest(event, repo)
 	Equals(t, models.PullRequest{
 		URL:        "https://gitlab.com/lkysow/atlantis-example/merge_requests/8",
 		Author:     "lkysow",
@@ -188,11 +214,12 @@ func TestParseGitlabMergeRequest(t *testing.T) {
 		HeadCommit: "0b4ac85ea3063ad5f2974d10cd68dd1f937aaac2",
 		Branch:     "abc",
 		State:      models.Open,
+		BaseRepo:   repo,
 	}, pull)
 
 	t.Log("If the state is closed, should set field correctly.")
 	event.State = "closed"
-	pull = parser.ParseGitlabMergeRequest(event)
+	pull = parser.ParseGitlabMergeRequest(event, repo)
 	Equals(t, models.Closed, pull.State)
 }
 
@@ -209,7 +236,10 @@ func TestParseGitlabMergeCommentEvent(t *testing.T) {
 		SanitizedCloneURL: "https://example.com/gitlabhq/gitlab-test.git",
 		Owner:             "gitlabhq",
 		CloneURL:          "https://gitlab-user:gitlab-token@example.com/gitlabhq/gitlab-test.git",
-		Hostname:          "example.com",
+		VCSHost: models.VCSHost{
+			Hostname: "example.com",
+			Type:     models.Gitlab,
+		},
 	}, baseRepo)
 	Equals(t, models.Repo{
 		FullName:          "gitlab-org/gitlab-test",
@@ -217,7 +247,10 @@ func TestParseGitlabMergeCommentEvent(t *testing.T) {
 		SanitizedCloneURL: "https://example.com/gitlab-org/gitlab-test.git",
 		Owner:             "gitlab-org",
 		CloneURL:          "https://gitlab-user:gitlab-token@example.com/gitlab-org/gitlab-test.git",
-		Hostname:          "example.com",
+		VCSHost: models.VCSHost{
+			Hostname: "example.com",
+			Type:     models.Gitlab,
+		},
 	}, headRepo)
 	Equals(t, models.User{
 		Username: "root",