+++ server/events_controller_test.go
@@ -31,7 +31,6 @@ import (
 	emocks "github.com/runatlantis/atlantis/server/events/mocks"
 	"github.com/runatlantis/atlantis/server/events/mocks/matchers"
 	"github.com/runatlantis/atlantis/server/events/models"
-	"github.com/runatlantis/atlantis/server/events/vcs"
 	vcsmocks "github.com/runatlantis/atlantis/server/events/vcs/mocks"
 	"github.com/runatlantis/atlantis/server/logging"
 	"github.com/runatlantis/atlantis/server/mocks"
@@ -42,13 +41,13 @@ const githubHeader = "X-Github-Event"
 const gitlabHeader = "X-Gitlab-Event"
 
 var secret = []byte("secret")
-var eventsReq *http.Request
 
 func TestPost_NotGithubOrGitlab(t *testing.T) {
 	t.Log("when the request is not for gitlab or github a 400 is returned")
 	e, _, _, _, _, _, _, _ := setup(t)
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	e.Post(w, req)
 	responseContains(t, w, http.StatusBadRequest, "Ignoring request")
 }
 
@@ -56,9 +55,10 @@ func TestPost_UnsupportedVCSGithub(t *testing.T) {
 	t.Log("when the request is for an unsupported vcs a 400 is returned")
 	e, _, _, _, _, _, _, _ := setup(t)
 	e.SupportedVCSHosts = nil
-	eventsReq.Header.Set(githubHeader, "value")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "value")
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusBadRequest, "Ignoring request since not configured to support GitHub")
 }
 
@@ -66,9 +66,10 @@ func TestPost_UnsupportedVCSGitlab(t *testing.T) {
 	t.Log("when the request is for an unsupported vcs a 400 is returned")
 	e, _, _, _, _, _, _, _ := setup(t)
 	e.SupportedVCSHosts = nil
-	eventsReq.Header.Set(gitlabHeader, "value")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(gitlabHeader, "value")
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusBadRequest, "Ignoring request since not configured to support GitLab")
 }
 
@@ -76,9 +77,10 @@ func TestPost_InvalidGithubSecret(t *testing.T) {
 	t.Log("when the github payload can't be validated a 400 is returned")
 	e, v, _, _, _, _, _, _ := setup(t)
 	w := httptest.NewRecorder()
-	eventsReq.Header.Set(githubHeader, "value")
-	When(v.Validate(eventsReq, secret)).ThenReturn(nil, errors.New("err"))
-	e.Post(w, eventsReq)
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "value")
+	When(v.Validate(req, secret)).ThenReturn(nil, errors.New("err"))
+	e.Post(w, req)
 	responseContains(t, w, http.StatusBadRequest, "err")
 }
 
@@ -86,9 +88,10 @@ func TestPost_InvalidGitlabSecret(t *testing.T) {
 	t.Log("when the gitlab payload can't be validated a 400 is returned")
 	e, _, gl, _, _, _, _, _ := setup(t)
 	w := httptest.NewRecorder()
-	eventsReq.Header.Set(gitlabHeader, "value")
-	When(gl.Validate(eventsReq, secret)).ThenReturn(nil, errors.New("err"))
-	e.Post(w, eventsReq)
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(gitlabHeader, "value")
+	When(gl.Validate(req, secret)).ThenReturn(nil, errors.New("err"))
+	e.Post(w, req)
 	responseContains(t, w, http.StatusBadRequest, "err")
 }
 
@@ -96,9 +99,10 @@ func TestPost_UnsupportedGithubEvent(t *testing.T) {
 	t.Log("when the event type is an unsupported github event we ignore it")
 	e, v, _, _, _, _, _, _ := setup(t)
 	w := httptest.NewRecorder()
-	eventsReq.Header.Set(githubHeader, "value")
-	When(v.Validate(eventsReq, nil)).ThenReturn([]byte(`{"not an event": ""}`), nil)
-	e.Post(w, eventsReq)
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "value")
+	When(v.Validate(req, nil)).ThenReturn([]byte(`{"not an event": ""}`), nil)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusOK, "Ignoring unsupported event")
 }
 
@@ -106,57 +110,62 @@ func TestPost_UnsupportedGitlabEvent(t *testing.T) {
 	t.Log("when the event type is an unsupported gitlab event we ignore it")
 	e, _, gl, _, _, _, _, _ := setup(t)
 	w := httptest.NewRecorder()
-	eventsReq.Header.Set(gitlabHeader, "value")
-	When(gl.Validate(eventsReq, secret)).ThenReturn([]byte(`{"not an event": ""}`), nil)
-	e.Post(w, eventsReq)
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(gitlabHeader, "value")
+	When(gl.Validate(req, secret)).ThenReturn([]byte(`{"not an event": ""}`), nil)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusOK, "Ignoring unsupported event")
 }
 
 func TestPost_GithubCommentNotCreated(t *testing.T) {
 	t.Log("when the event is a github comment but it's not a created event we ignore it")
 	e, v, _, _, _, _, _, _ := setup(t)
-	eventsReq.Header.Set(githubHeader, "issue_comment")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "issue_comment")
 	// comment action is deleted, not created
 	event := `{"action": "deleted"}`
-	When(v.Validate(eventsReq, secret)).ThenReturn([]byte(event), nil)
+	When(v.Validate(req, secret)).ThenReturn([]byte(event), nil)
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusOK, "Ignoring comment event since action was not created")
 }
 
 func TestPost_GithubInvalidComment(t *testing.T) {
 	t.Log("when the event is a github comment without all expected data we return a 400")
 	e, v, _, p, _, _, _, _ := setup(t)
-	eventsReq.Header.Set(githubHeader, "issue_comment")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "issue_comment")
 	event := `{"action": "created"}`
-	When(v.Validate(eventsReq, secret)).ThenReturn([]byte(event), nil)
+	When(v.Validate(req, secret)).ThenReturn([]byte(event), nil)
 	When(p.ParseGithubIssueCommentEvent(matchers.AnyPtrToGithubIssueCommentEvent())).ThenReturn(models.Repo{}, models.User{}, 1, errors.New("err"))
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusBadRequest, "Failed parsing event")
 }
 
 func TestPost_GitlabCommentInvalidCommand(t *testing.T) {
 	t.Log("when the event is a gitlab comment with an invalid command we ignore it")
 	e, _, gl, _, _, _, _, cp := setup(t)
-	eventsReq.Header.Set(gitlabHeader, "value")
-	When(gl.Validate(eventsReq, secret)).ThenReturn(gitlab.MergeCommentEvent{}, nil)
-	When(cp.Parse("", vcs.Gitlab)).ThenReturn(events.CommentParseResult{Ignore: true})
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(gitlabHeader, "value")
+	When(gl.Validate(req, secret)).ThenReturn(gitlab.MergeCommentEvent{}, nil)
+	When(cp.Parse("", models.Gitlab)).ThenReturn(events.CommentParseResult{Ignore: true})
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusOK, "Ignoring non-command comment: \"\"")
 }
 
 func TestPost_GithubCommentInvalidCommand(t *testing.T) {
 	t.Log("when the event is a github comment with an invalid command we ignore it")
 	e, v, _, p, _, _, _, cp := setup(t)
-	eventsReq.Header.Set(githubHeader, "issue_comment")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "issue_comment")
 	event := `{"action": "created"}`
-	When(v.Validate(eventsReq, secret)).ThenReturn([]byte(event), nil)
+	When(v.Validate(req, secret)).ThenReturn([]byte(event), nil)
 	When(p.ParseGithubIssueCommentEvent(matchers.AnyPtrToGithubIssueCommentEvent())).ThenReturn(models.Repo{}, models.User{}, 1, nil)
-	When(cp.Parse("", vcs.Github)).ThenReturn(events.CommentParseResult{Ignore: true})
+	When(cp.Parse("", models.Github)).ThenReturn(events.CommentParseResult{Ignore: true})
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusOK, "Ignoring non-command comment: \"\"")
 }
 
@@ -169,7 +178,7 @@ func TestPost_GitlabCommentNotWhitelisted(t *testing.T) {
 		CommentParser:       &events.CommentParser{},
 		GitlabRequestParser: &server.DefaultGitlabRequestParser{},
 		Parser:              &events.EventParser{},
-		SupportedVCSHosts:   []vcs.Host{vcs.Gitlab},
+		SupportedVCSHosts:   []models.VCSHostType{models.Gitlab},
 		RepoWhitelist:       &events.RepoWhitelist{},
 		VCSClient:           vcsClient,
 	}
@@ -184,8 +193,8 @@ func TestPost_GitlabCommentNotWhitelisted(t *testing.T) {
 	body, _ := ioutil.ReadAll(w.Result().Body)
 	exp := "Repo not whitelisted"
 	Assert(t, strings.Contains(string(body), exp), "exp %q to be contained in %q", exp, string(body))
-	expRepo, _ := models.NewRepo("gitlabhq/gitlab-test", "https://example.com/gitlabhq/gitlab-test.git", "", "")
-	vcsClient.VerifyWasCalledOnce().CreateComment(expRepo, 1, "```\nError: This repo is not whitelisted for Atlantis.\n```", vcs.Gitlab)
+	expRepo, _ := models.NewRepo(models.Gitlab, "gitlabhq/gitlab-test", "https://example.com/gitlabhq/gitlab-test.git", "", "")
+	vcsClient.VerifyWasCalledOnce().CreateComment(expRepo, 1, "```\nError: This repo is not whitelisted for Atlantis.\n```")
 }
 
 func TestPost_GithubCommentNotWhitelisted(t *testing.T) {
@@ -197,7 +206,7 @@ func TestPost_GithubCommentNotWhitelisted(t *testing.T) {
 		GithubRequestValidator: &server.DefaultGithubRequestValidator{},
 		CommentParser:          &events.CommentParser{},
 		Parser:                 &events.EventParser{},
-		SupportedVCSHosts:      []vcs.Host{vcs.Github},
+		SupportedVCSHosts:      []models.VCSHostType{models.Github},
 		RepoWhitelist:          &events.RepoWhitelist{},
 		VCSClient:              vcsClient,
 	}
@@ -213,120 +222,128 @@ func TestPost_GithubCommentNotWhitelisted(t *testing.T) {
 	body, _ := ioutil.ReadAll(w.Result().Body)
 	exp := "Repo not whitelisted"
 	Assert(t, strings.Contains(string(body), exp), "exp %q to be contained in %q", exp, string(body))
-	expRepo, _ := models.NewRepo("baxterthehacker/public-repo", "https://github.com/baxterthehacker/public-repo.git", "", "")
-	vcsClient.VerifyWasCalledOnce().CreateComment(expRepo, 2, "```\nError: This repo is not whitelisted for Atlantis.\n```", vcs.Github)
+	expRepo, _ := models.NewRepo(models.Github, "baxterthehacker/public-repo", "https://github.com/baxterthehacker/public-repo.git", "", "")
+	vcsClient.VerifyWasCalledOnce().CreateComment(expRepo, 2, "```\nError: This repo is not whitelisted for Atlantis.\n```")
 }
 
 func TestPost_GitlabCommentResponse(t *testing.T) {
 	// When the event is a gitlab comment that warrants a comment response we comment back.
 	e, _, gl, _, _, _, vcsClient, cp := setup(t)
-	eventsReq.Header.Set(gitlabHeader, "value")
-	When(gl.Validate(eventsReq, secret)).ThenReturn(gitlab.MergeCommentEvent{}, nil)
-	When(cp.Parse("", vcs.Gitlab)).ThenReturn(events.CommentParseResult{CommentResponse: "a comment"})
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(gitlabHeader, "value")
+	When(gl.Validate(req, secret)).ThenReturn(gitlab.MergeCommentEvent{}, nil)
+	When(cp.Parse("", models.Gitlab)).ThenReturn(events.CommentParseResult{CommentResponse: "a comment"})
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
-	vcsClient.VerifyWasCalledOnce().CreateComment(models.Repo{}, 0, "a comment", vcs.Gitlab)
+	e.Post(w, req)
+	vcsClient.VerifyWasCalledOnce().CreateComment(models.Repo{}, 0, "a comment")
 	responseContains(t, w, http.StatusOK, "Commenting back on pull request")
 }
 
 func TestPost_GithubCommentResponse(t *testing.T) {
 	t.Log("when the event is a github comment that warrants a comment response we comment back")
 	e, v, _, p, _, _, vcsClient, cp := setup(t)
-	eventsReq.Header.Set(githubHeader, "issue_comment")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "issue_comment")
 	event := `{"action": "created"}`
-	When(v.Validate(eventsReq, secret)).ThenReturn([]byte(event), nil)
+	When(v.Validate(req, secret)).ThenReturn([]byte(event), nil)
 	baseRepo := models.Repo{}
 	user := models.User{}
 	When(p.ParseGithubIssueCommentEvent(matchers.AnyPtrToGithubIssueCommentEvent())).ThenReturn(baseRepo, user, 1, nil)
-	When(cp.Parse("", vcs.Github)).ThenReturn(events.CommentParseResult{CommentResponse: "a comment"})
+	When(cp.Parse("", models.Github)).ThenReturn(events.CommentParseResult{CommentResponse: "a comment"})
 	w := httptest.NewRecorder()
 
-	e.Post(w, eventsReq)
-	vcsClient.VerifyWasCalledOnce().CreateComment(baseRepo, 1, "a comment", vcs.Github)
+	e.Post(w, req)
+	vcsClient.VerifyWasCalledOnce().CreateComment(baseRepo, 1, "a comment")
 	responseContains(t, w, http.StatusOK, "Commenting back on pull request")
 }
 
 func TestPost_GitlabCommentSuccess(t *testing.T) {
 	t.Log("when the event is a gitlab comment with a valid command we call the command handler")
 	e, _, gl, _, cr, _, _, _ := setup(t)
-	eventsReq.Header.Set(gitlabHeader, "value")
-	When(gl.Validate(eventsReq, secret)).ThenReturn(gitlab.MergeCommentEvent{}, nil)
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(gitlabHeader, "value")
+	When(gl.Validate(req, secret)).ThenReturn(gitlab.MergeCommentEvent{}, nil)
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusOK, "Processing...")
 
 	// wait for 200ms so goroutine is called
 	time.Sleep(200 * time.Millisecond)
-	cr.VerifyWasCalledOnce().ExecuteCommand(models.Repo{}, models.Repo{}, models.User{}, 0, nil, vcs.Gitlab)
+	cr.VerifyWasCalledOnce().ExecuteCommand(models.Repo{}, models.Repo{}, models.User{}, 0, nil)
 }
 
 func TestPost_GithubCommentSuccess(t *testing.T) {
 	t.Log("when the event is a github comment with a valid command we call the command handler")
 	e, v, _, p, cr, _, _, cp := setup(t)
-	eventsReq.Header.Set(githubHeader, "issue_comment")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "issue_comment")
 	event := `{"action": "created"}`
-	When(v.Validate(eventsReq, secret)).ThenReturn([]byte(event), nil)
+	When(v.Validate(req, secret)).ThenReturn([]byte(event), nil)
 	baseRepo := models.Repo{}
 	user := models.User{}
 	cmd := events.Command{}
 	When(p.ParseGithubIssueCommentEvent(matchers.AnyPtrToGithubIssueCommentEvent())).ThenReturn(baseRepo, user, 1, nil)
-	When(cp.Parse("", vcs.Github)).ThenReturn(events.CommentParseResult{Command: &cmd})
+	When(cp.Parse("", models.Github)).ThenReturn(events.CommentParseResult{Command: &cmd})
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusOK, "Processing...")
 
 	// wait for 200ms so goroutine is called
 	time.Sleep(200 * time.Millisecond)
-	cr.VerifyWasCalledOnce().ExecuteCommand(baseRepo, baseRepo, user, 1, &cmd, vcs.Github)
+	cr.VerifyWasCalledOnce().ExecuteCommand(baseRepo, baseRepo, user, 1, &cmd)
 }
 
 func TestPost_GithubPullRequestNotClosed(t *testing.T) {
 	t.Log("when the event is a github pull reuqest but it's not a closed event we ignore it")
 	e, v, _, _, _, _, _, _ := setup(t)
-	eventsReq.Header.Set(githubHeader, "pull_request")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "pull_request")
 	event := `{"action": "opened"}`
-	When(v.Validate(eventsReq, secret)).ThenReturn([]byte(event), nil)
+	When(v.Validate(req, secret)).ThenReturn([]byte(event), nil)
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusOK, "Ignoring opened pull request event")
 }
 
 func TestPost_GitlabMergeRequestNotClosed(t *testing.T) {
 	t.Log("when the event is a gitlab merge request but it's not a closed event we ignore it")
 	e, _, gl, p, _, _, _, _ := setup(t)
-	eventsReq.Header.Set(gitlabHeader, "value")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(gitlabHeader, "value")
 	event := gitlab.MergeEvent{}
-	When(gl.Validate(eventsReq, secret)).ThenReturn(event, nil)
+	When(gl.Validate(req, secret)).ThenReturn(event, nil)
 	When(p.ParseGitlabMergeEvent(event)).ThenReturn(models.PullRequest{State: models.Open}, models.Repo{}, nil)
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusOK, "Ignoring opened pull request event")
 }
 
 func TestPost_GithubPullRequestInvalid(t *testing.T) {
 	t.Log("when the event is a github pull request with invalid data we return a 400")
 	e, v, _, p, _, _, _, _ := setup(t)
-	eventsReq.Header.Set(githubHeader, "pull_request")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "pull_request")
 
 	event := `{"action": "closed"}`
-	When(v.Validate(eventsReq, secret)).ThenReturn([]byte(event), nil)
+	When(v.Validate(req, secret)).ThenReturn([]byte(event), nil)
 	When(p.ParseGithubPull(matchers.AnyPtrToGithubPullRequest())).ThenReturn(models.PullRequest{}, models.Repo{}, errors.New("err"))
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusBadRequest, "Error parsing pull data: err")
 }
 
 func TestPost_GithubPullRequestInvalidRepo(t *testing.T) {
 	t.Log("when the event is a github pull request with invalid repo data we return a 400")
 	e, v, _, p, _, _, _, _ := setup(t)
-	eventsReq.Header.Set(githubHeader, "pull_request")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "pull_request")
 
 	event := `{"action": "closed"}`
-	When(v.Validate(eventsReq, secret)).ThenReturn([]byte(event), nil)
+	When(v.Validate(req, secret)).ThenReturn([]byte(event), nil)
 	When(p.ParseGithubPull(matchers.AnyPtrToGithubPullRequest())).ThenReturn(models.PullRequest{}, models.Repo{}, nil)
 	When(p.ParseGithubRepo(matchers.AnyPtrToGithubRepository())).ThenReturn(models.Repo{}, errors.New("err"))
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusBadRequest, "Error parsing repo data: err")
 }
 
@@ -334,14 +351,15 @@ func TestPost_GithubPullRequestNotWhitelisted(t *testing.T) {
 	t.Log("when the event is a github pull request to a non-whitelisted repo we return a 400")
 	e, v, _, p, _, _, _, _ := setup(t)
 	e.RepoWhitelist = &events.RepoWhitelist{Whitelist: "github.com/nevermatch"}
-	eventsReq.Header.Set(githubHeader, "pull_request")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "pull_request")
 
 	event := `{"action": "closed"}`
-	When(v.Validate(eventsReq, secret)).ThenReturn([]byte(event), nil)
+	When(v.Validate(req, secret)).ThenReturn([]byte(event), nil)
 	When(p.ParseGithubPull(matchers.AnyPtrToGithubPullRequest())).ThenReturn(models.PullRequest{}, models.Repo{}, nil)
 	When(p.ParseGithubRepo(matchers.AnyPtrToGithubRepository())).ThenReturn(models.Repo{}, nil)
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusForbidden, "Ignoring pull request event from non-whitelisted repo")
 }
 
@@ -349,69 +367,72 @@ func TestPost_GithubPullRequestErrCleaningPull(t *testing.T) {
 	t.Log("when the event is a pull request and we have an error calling CleanUpPull we return a 503")
 	RegisterMockTestingT(t)
 	e, v, _, p, _, c, _, _ := setup(t)
-	eventsReq.Header.Set(githubHeader, "pull_request")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "pull_request")
 
 	event := `{"action": "closed"}`
-	When(v.Validate(eventsReq, secret)).ThenReturn([]byte(event), nil)
+	When(v.Validate(req, secret)).ThenReturn([]byte(event), nil)
 	repo := models.Repo{}
 	pull := models.PullRequest{State: models.Closed}
 	When(p.ParseGithubPull(matchers.AnyPtrToGithubPullRequest())).ThenReturn(pull, repo, nil)
 	When(p.ParseGithubRepo(matchers.AnyPtrToGithubRepository())).ThenReturn(repo, nil)
-	When(c.CleanUpPull(repo, pull, vcs.Github)).ThenReturn(errors.New("cleanup err"))
+	When(c.CleanUpPull(repo, pull)).ThenReturn(errors.New("cleanup err"))
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusInternalServerError, "Error cleaning pull request: cleanup err")
 }
 
 func TestPost_GitlabMergeRequestErrCleaningPull(t *testing.T) {
 	t.Log("when the event is a gitlab merge request and an error occurs calling CleanUpPull we return a 503")
 	e, _, gl, p, _, c, _, _ := setup(t)
-	eventsReq.Header.Set(gitlabHeader, "value")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(gitlabHeader, "value")
 	event := gitlab.MergeEvent{}
-	When(gl.Validate(eventsReq, secret)).ThenReturn(event, nil)
+	When(gl.Validate(req, secret)).ThenReturn(event, nil)
 	repo := models.Repo{}
 	pullRequest := models.PullRequest{State: models.Closed}
 	When(p.ParseGitlabMergeEvent(event)).ThenReturn(pullRequest, repo, nil)
-	When(c.CleanUpPull(repo, pullRequest, vcs.Gitlab)).ThenReturn(errors.New("err"))
+	When(c.CleanUpPull(repo, pullRequest)).ThenReturn(errors.New("err"))
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusInternalServerError, "Error cleaning pull request: err")
 }
 
 func TestPost_GithubPullRequestSuccess(t *testing.T) {
 	t.Log("when the event is a pull request and everything works we return a 200")
 	e, v, _, p, _, c, _, _ := setup(t)
-	eventsReq.Header.Set(githubHeader, "pull_request")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(githubHeader, "pull_request")
 
 	event := `{"action": "closed"}`
-	When(v.Validate(eventsReq, secret)).ThenReturn([]byte(event), nil)
+	When(v.Validate(req, secret)).ThenReturn([]byte(event), nil)
 	repo := models.Repo{}
 	pull := models.PullRequest{State: models.Closed}
 	When(p.ParseGithubPull(matchers.AnyPtrToGithubPullRequest())).ThenReturn(pull, repo, nil)
 	When(p.ParseGithubRepo(matchers.AnyPtrToGithubRepository())).ThenReturn(repo, nil)
-	When(c.CleanUpPull(repo, pull, vcs.Github)).ThenReturn(nil)
+	When(c.CleanUpPull(repo, pull)).ThenReturn(nil)
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusOK, "Pull request cleaned successfully")
 }
 
 func TestPost_GitlabMergeRequestSuccess(t *testing.T) {
 	t.Log("when the event is a gitlab merge request and the cleanup works we return a 200")
 	e, _, gl, p, _, _, _, _ := setup(t)
-	eventsReq.Header.Set(gitlabHeader, "value")
+	req, _ := http.NewRequest("GET", "", bytes.NewBuffer(nil))
+	req.Header.Set(gitlabHeader, "value")
 	event := gitlab.MergeEvent{}
-	When(gl.Validate(eventsReq, secret)).ThenReturn(event, nil)
+	When(gl.Validate(req, secret)).ThenReturn(event, nil)
 	repo := models.Repo{}
 	pullRequest := models.PullRequest{State: models.Closed}
 	When(p.ParseGitlabMergeEvent(event)).ThenReturn(pullRequest, repo, nil)
 	w := httptest.NewRecorder()
-	e.Post(w, eventsReq)
+	e.Post(w, req)
 	responseContains(t, w, http.StatusOK, "Pull request cleaned successfully")
 }
 
 func setup(t *testing.T) (server.EventsController, *mocks.MockGithubRequestValidator, *mocks.MockGitlabRequestParser, *emocks.MockEventParsing, *emocks.MockCommandRunner, *emocks.MockPullCleaner, *vcsmocks.MockClientProxy, *emocks.MockCommentParsing) {
 	RegisterMockTestingT(t)
-	eventsReq, _ = http.NewRequest("GET", "", bytes.NewBuffer(nil))
 	v := mocks.NewMockGithubRequestValidator()
 	gl := mocks.NewMockGitlabRequestParser()
 	p := emocks.NewMockEventParsing()
@@ -427,7 +448,7 @@ func setup(t *testing.T) (server.EventsController, *mocks.MockGithubRequestValid
 		CommandRunner:          cr,
 		PullCleaner:            c,
 		GithubWebHookSecret:    secret,
-		SupportedVCSHosts:      []vcs.Host{vcs.Github, vcs.Gitlab},
+		SupportedVCSHosts:      []models.VCSHostType{models.Github, models.Gitlab},
 		GitlabWebHookSecret:    secret,
 		GitlabRequestParser:    gl,
 		RepoWhitelist: &events.RepoWhitelist{