+++ server/server.go
@@ -35,6 +35,7 @@ import (
 	"github.com/runatlantis/atlantis/server/events"
 	"github.com/runatlantis/atlantis/server/events/locking"
 	"github.com/runatlantis/atlantis/server/events/locking/boltdb"
+	"github.com/runatlantis/atlantis/server/events/models"
 	"github.com/runatlantis/atlantis/server/events/run"
 	"github.com/runatlantis/atlantis/server/events/terraform"
 	"github.com/runatlantis/atlantis/server/events/vcs"
@@ -57,6 +58,7 @@ type Server struct {
 	Locker             locking.Locker
 	AtlantisURL        string
 	EventsController   *EventsController
+	LocksController    *LocksController
 	IndexTemplate      TemplateWriter
 	LockDetailTemplate TemplateWriter
 	SSLCertFile        string
@@ -115,11 +117,11 @@ type WebhookConfig struct {
 // its dependencies an error will be returned. This is like the main() function
 // for the server CLI command because it injects all the dependencies.
 func NewServer(userConfig UserConfig, config Config) (*Server, error) {
-	var supportedVCSHosts []vcs.Host
+	var supportedVCSHosts []models.VCSHostType
 	var githubClient *vcs.GithubClient
 	var gitlabClient *vcs.GitlabClient
 	if userConfig.GithubUser != "" {
-		supportedVCSHosts = append(supportedVCSHosts, vcs.Github)
+		supportedVCSHosts = append(supportedVCSHosts, models.Github)
 		var err error
 		githubClient, err = vcs.NewGithubClient(userConfig.GithubHostname, userConfig.GithubUser, userConfig.GithubToken)
 		if err != nil {
@@ -127,7 +129,7 @@ func NewServer(userConfig UserConfig, config Config) (*Server, error) {
 		}
 	}
 	if userConfig.GitlabUser != "" {
-		supportedVCSHosts = append(supportedVCSHosts, vcs.Gitlab)
+		supportedVCSHosts = append(supportedVCSHosts, models.Gitlab)
 		gitlabClient = &vcs.GitlabClient{
 			Client: gitlab.NewClient(nil, userConfig.GitlabToken),
 		}
@@ -242,6 +244,13 @@ func NewServer(userConfig UserConfig, config Config) (*Server, error) {
 	repoWhitelist := &events.RepoWhitelist{
 		Whitelist: userConfig.RepoWhitelist,
 	}
+	locksController := &LocksController{
+		AtlantisVersion:    config.AtlantisVersion,
+		Locker:             lockingClient,
+		Logger:             logger,
+		VCSClient:          vcsClient,
+		LockDetailTemplate: lockTemplate,
+	}
 	eventsController := &EventsController{
 		CommandRunner:          commandHandler,
 		PullCleaner:            pullClosedExecutor,
@@ -266,6 +275,7 @@ func NewServer(userConfig UserConfig, config Config) (*Server, error) {
 		Locker:             lockingClient,
 		AtlantisURL:        userConfig.AtlantisURL,
 		EventsController:   eventsController,
+		LocksController:    locksController,
 		IndexTemplate:      indexTemplate,
 		LockDetailTemplate: lockTemplate,
 		SSLKeyFile:         userConfig.SSLKeyFile,
@@ -279,9 +289,9 @@ func (s *Server) Start() error {
 		return r.URL.Path == "/" || r.URL.Path == "/index.html"
 	})
 	s.Router.PathPrefix("/static/").Handler(http.FileServer(&assetfs.AssetFS{Asset: static.Asset, AssetDir: static.AssetDir, AssetInfo: static.AssetInfo}))
-	s.Router.HandleFunc("/events", s.postEvents).Methods("POST")
-	s.Router.HandleFunc("/locks", s.DeleteLockRoute).Methods("DELETE").Queries("id", "{id:.*}")
-	lockRoute := s.Router.HandleFunc("/lock", s.GetLockRoute).Methods("GET").Queries("id", "{id}").Name(LockRouteName)
+	s.Router.HandleFunc("/events", s.EventsController.Post).Methods("POST")
+	s.Router.HandleFunc("/locks", s.LocksController.DeleteLock).Methods("DELETE").Queries("id", "{id:.*}")
+	lockRoute := s.Router.HandleFunc("/lock", s.LocksController.GetLock).Methods("GET").Queries("id", "{id}").Name(LockRouteName)
 	// function that planExecutor can use to construct detail view url
 	// injecting this here because this is the earliest routes are created
 	s.CommandHandler.SetLockURL(func(lockID string) string {
@@ -352,99 +362,4 @@ func (s *Server) Index(w http.ResponseWriter, _ *http.Request) {
 		Locks:           lockResults,
 		AtlantisVersion: s.AtlantisVersion,
 	})
-}
-
-// GetLockRoute is the GET /locks/{id} route. It renders the lock detail view.
-func (s *Server) GetLockRoute(w http.ResponseWriter, r *http.Request) {
-	id, ok := mux.Vars(r)["id"]
-	if !ok {
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprint(w, "No lock id in request")
-		return
-	}
-	s.GetLock(w, r, id)
-}
-
-// GetLock handles a lock detail page view. getLockRoute is expected to
-// be called before. This function was extracted to make it testable.
-func (s *Server) GetLock(w http.ResponseWriter, _ *http.Request, id string) {
-	idUnencoded, err := url.QueryUnescape(id)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprint(w, "Invalid lock id")
-		return
-	}
-
-	lock, err := s.Locker.GetLock(idUnencoded)
-	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprint(w, err.Error())
-		return
-	}
-	if lock == nil {
-		w.WriteHeader(http.StatusNotFound)
-		fmt.Fprint(w, "No lock found at that id")
-		return
-	}
-
-	// Extract the repo owner and repo name.
-	repo := strings.Split(lock.Project.RepoFullName, "/")
-
-	l := LockDetailData{
-		LockKeyEncoded:  id,
-		LockKey:         idUnencoded,
-		RepoOwner:       repo[0],
-		RepoName:        repo[1],
-		PullRequestLink: lock.Pull.URL,
-		LockedBy:        lock.Pull.Author,
-		Workspace:       lock.Workspace,
-		AtlantisVersion: s.AtlantisVersion,
-	}
-
-	s.LockDetailTemplate.Execute(w, l) // nolint: errcheck
-}
-
-// DeleteLockRoute handles deleting the lock at id.
-func (s *Server) DeleteLockRoute(w http.ResponseWriter, r *http.Request) {
-	id, ok := mux.Vars(r)["id"]
-	if !ok || id == "" {
-		s.respond(w, logging.Warn, http.StatusBadRequest, "No lock id in request")
-		return
-	}
-	s.DeleteLock(w, r, id)
-}
-
-// DeleteLock deletes the lock. DeleteLockRoute should be called first.
-// This method is split out to make this route testable.
-func (s *Server) DeleteLock(w http.ResponseWriter, _ *http.Request, id string) {
-	idUnencoded, err := url.PathUnescape(id)
-	if err != nil {
-		s.respond(w, logging.Warn, http.StatusBadRequest, "Invalid lock id: %s", err)
-		return
-	}
-	lock, err := s.Locker.Unlock(idUnencoded)
-	if err != nil {
-		s.respond(w, logging.Error, http.StatusInternalServerError, "Failed to delete lock %s: %s", idUnencoded, err)
-		return
-	}
-	if lock == nil {
-		s.respond(w, logging.Warn, http.StatusNotFound, "No lock found at that id", idUnencoded)
-		return
-	}
-	s.respond(w, logging.Info, http.StatusOK, "Deleted lock id %s", idUnencoded)
-}
-
-// postEvents handles POST requests to our /events endpoint. These should be
-// VCS webhook requests.
-func (s *Server) postEvents(w http.ResponseWriter, r *http.Request) {
-	s.EventsController.Post(w, r)
-}
-
-// respond is a helper function to respond and log the response. lvl is the log
-// level to log at, code is the HTTP response code.
-func (s *Server) respond(w http.ResponseWriter, lvl logging.LogLevel, code int, format string, args ...interface{}) {
-	response := fmt.Sprintf(format, args...)
-	s.Logger.Log(lvl, response)
-	w.WriteHeader(code)
-	fmt.Fprintln(w, response)
 }